Create fail 2
There is an internal library called arrays, which can perform multiple operations on arrays.
https://docs.oracle.com/javase/7/docs/api/java/util/Arrays.html

Find the function that can sort an array, and use it to sort the array given in the question. 
(Hint:  Use the internal library sort() to sort a given array. For e.g.,  sort(long[] a) sorts the specified array in ascending numerical order.) 

Input:
No Input Required